Server is running again.
