+++ backend/server.js
@@ -11,6 +11,7 @@ const bodyParser = require('body-parser');
 // import routes
 const users = require("./routes/api/user");
 const classes = require("./routes/api/class");
+// const student = require("./routes/api2/student");
 const billing = require("./routes/api/payment");
 const githubData = require("./data/githubData");
 
@@ -26,35 +27,35 @@ app.use(express.json());
 app.use(helmet());
 app.use(cors());
 app.use(fileUpload());
-
+app.use(require('sanitize').middleware);
 
 
 
 // // ****START STRIPE****
 
-// const CORS_WHITELIST = require('./billing/frontend');
+const CORS_WHITELIST = require('./billing/frontend');
+// const WHITELIST = [CORS_WHITELIST, 'http://localhost:3000/']
+const corsOptions = {
+    origin: (origin, callback) =>
+        (CORS_WHITELIST.indexOf(origin) !== -1)
+            ? callback(null,true)
+            : callback(new Error('Not allowed by CORS'))
+};
 
-// const corsOptions = {
-//   origin: (origin, callback) =>
-//     (CORS_WHITELIST.indexOf(origin) !== -1)
-//       ? callback(null,true)
-//       : callback(new Error('Not allowed by CORS'))
-// };
+const configureServer = app => {
+    app.use(cors(corsOptions));
+    app.use(bodyParser.json());
+};
 
-// const configureServer = app => {
-//   app.use(cors(corsOptions));
-//   app.use(bodyParser.json());
-// };
+const paymentApi = require('./billing/payment');
 
-// const paymentApi = require('./billing/payment');
-
-// const configureRoutes = app => {
-//   paymentApi(app);
-// };
+const configureRoutes = app => {
+    paymentApi(app);
+};
 
 
-// configureServer(app);
-// configureRoutes(app);
+configureServer(app);
+configureRoutes(app);
 
 // // ****END STRIPE****
 
@@ -67,30 +68,31 @@ app.use(express.static(path.join(__dirname, "../leaderboard-frontend/build")));
 // Connect MongoDB
 const db = process.env.MONGO_URI;
 mongoose
-  .connect(
-    db,
-    { useNewUrlParser: true }
-  )
-  .then(() => console.log("=== Connected to MongoDB ===\n"))
-  .catch(err => console.log(err));
+    .connect(
+        db,
+        { useNewUrlParser: true }
+    )
+    .then(() => console.log("=== Connected to MongoDB ===\n"))
+    .catch(err => console.log(err));
 
 // Set up passport middleware
 app.use(passport.initialize());
 require("./authentication/passport")(passport);
 
 // Connect routes
-app.use("/api/users", users);
+app.use("/api/users", cors(corsOptions), users);
 app.use(
-  "/api/classes",
-  passport.authenticate("jwt", { session: false }),
-  classes
+    "/api/classes",
+    cors(corsOptions),
+    passport.authenticate("jwt", { session: false }),
+    classes
 );
-// app.use("/api/data", githubData);
-app.use("/api/billing", billing);
+app.use("/api/data", cors(corsOptions), githubData);
+app.use("/api/billing",cors(corsOptions), billing);
 
 // CSV routes
-//app.get("/template");
-app.post("/create-edit", upload.post);
+app.get("/template", cors(corsOptions),);
+app.post("/create-edit", cors(corsOptions), upload.post);
 
 const port = process.env.PORT || 4000;
